solver: split platform specific code
