+++ solver/socketprovider.go
@@ -3,11 +3,8 @@ package solver
 import (
 	"context"
 	"fmt"
-	"net"
 	"strings"
-	"time"
 
-	"github.com/Microsoft/go-winio"
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/session/sshforward"
 	"google.golang.org/grpc"
@@ -39,20 +36,6 @@ func (sp *SocketProvider) CheckAgent(ctx context.Context, req *sshforward.CheckA
 		return &sshforward.CheckAgentResponse{}, fmt.Errorf("invalid socket forward key %s", id)
 	}
 	return &sshforward.CheckAgentResponse{}, nil
-}
-
-func dialStream(id string) (net.Conn, error) {
-	switch {
-	case strings.HasPrefix(id, unixPrefix):
-		id = strings.TrimPrefix(id, unixPrefix)
-		return net.DialTimeout("unix", id, time.Second)
-	case strings.HasPrefix(id, npipePrefix):
-		id = strings.TrimPrefix(id, npipePrefix)
-		dur := time.Second
-		return winio.DialPipe(id, &dur)
-	default:
-		return nil, fmt.Errorf("invalid socket forward key %s", id)
-	}
 }
 
 func (sp *SocketProvider) ForwardAgent(stream sshforward.SSH_ForwardAgentServer) error {
